+++ lib/src/features/test/test_bloc.dart
@@ -1,4 +1,3 @@
-import 'dart:async';
 
 import 'package:bloc/bloc.dart';
 import 'package:meta/meta.dart';
@@ -8,8 +7,6 @@ part 'test_state.dart';
 
 class TestBloc extends Bloc<TestEvent, TestState> {
   TestBloc() : super(TestInitial()) {
-    on<TestEvent>((event, emit) {
-      // TODO: implement event handler
-    });
+    on<TestEvent>((event, emit) {});
   }
 }