+++ CHANGELOG.md
@@ -2,7 +2,16 @@
 - Initial public release
 
 ## [0.0.2] - 1 FEB 2022
-- Code optimized
+- Code optimizedN
 
 ## [0.0.3] - 1 FEB 2022
 - Added GlobalResponseModel & code refactor
+
+## [0.0.4] - 2 FEB 2022
+- Removed android and ios folder dependency & optimized
+
+## [0.0.5] - 2 FEB 2022
+- Resolved shared preferences dependency issue on session manager
+
+## [0.0.6] - 2 FEB 2022
+- Downgraded bloc to version 7.x.x due to major changes in version 8.x.x